ActivityDetail: define in storage, re-export from shared/types
The producer (storage layer) should own the domain type; shared/types
just re-exports it for the renderer/IPC contract.
