+++ src/main/storage/types.ts
@@ -21,4 +21,13 @@ export interface ActivitySummary {
   summary: string
 }
 
-export type { ActivityDetail } from '../../shared/types'
+/** Activity with window context but without heavy ocr_text and vector fields. */
+export interface ActivityDetail {
+  id: string
+  startTimestamp: number
+  endTimestamp: number
+  appName: string
+  windowTitle: string
+  tld: string | null
+  summary: string
+}
